+++ kube/src/api/dynamic.rs
@@ -1,0 +1,26 @@
+#[cfg(test)]
+mod test {
+    #[cfg(feature = "derive")]
+    #[tokio::test]
+    #[ignore] // circle has no kubeconfig
+    async fn convenient_custom_resource() {
+        use crate as kube; // derive macro needs kube in scope
+        use kube::{Api, Client, CustomResource, api::{GroupVersionKind, DynamicObject, ApiResource}};
+        use schemars::JsonSchema;
+        use serde::{Deserialize, Serialize};
+        #[derive(Clone, Debug, CustomResource, Deserialize, Serialize, JsonSchema)]
+        #[kube(group = "clux.dev", version = "v1", kind = "Foo", namespaced)]
+        struct FooSpec {
+            foo: String,
+        }
+        let client = Client::try_default().await.unwrap();
+
+        let gvk = GroupVersionKind::gvk("clux.dev", "v1", "Foo");
+        let api_resource = ApiResource::from_gvk(&gvk);
+        let a1: Api<DynamicObject> = Api::namespaced_with(client.clone(), "myns", &api_resource);
+        let a2: Api<Foo> = Api::namespaced(client.clone(), "myns");
+
+        // make sure they return the same url_path through their impls
+        assert_eq!(a1.request.url_path, a2.request.url_path);
+    }
+}
